change terminal to be the default
