+++ .gitignore
@@ -1,7 +1,5 @@
 clue
 clue.exe
-clue-legacy
-clue-legacy.exe
 test
 test.exe
 .vscode/configurationCache.log
