--- draw.h
@@ -1,11 +1,0 @@
-#include "game.h"
-#include "items.h"
-#include "character.h"
-#include "rooms.h"
-
-#ifndef DRAW_H
-#define DRA_H
-
-void DrawMap(struct Room* map[9]);
-
-#endif
